+++ R/schedule.R
@@ -13,26 +13,26 @@ ypr_schedule <- function(population) {
 
   schedule <- with(population, {
     t <- tR:tmax
-    n <- length(t)
+    nt <- length(t)
     L <- Linf * (1 - exp(-k * (t - t0)))
     L[L < 0] <- 0
     W <- Wa * L^Wb
     E <- fa * W^fb
     S <- exp(log(L / Linf) * Sp) / (exp(log(Ls / Linf) * Sp) + exp(log(L / Linf) * Sp)) * es
-    N <- ypr_inst2inter(M * L^Mb)
-    N[1] <- 1 - (1 - N[1]) * (1 - A0)
+    N <- rep(n, nt)
+    N[L >= Ln] <- nL
     N <- 1 - ((1 - N) * (1 - S * Sm))
     V <- exp(log(L / Linf) * Vp) / (exp(log(Lv / Linf) * Vp) + exp(log(L / Linf) * Vp))
     C <- pi * V
-    R <- rep(1 - rho, n)
+    R <- rep(1 - rho, nt)
     R[L < Llo | L > Lup] <- Nc
     U <- C * R + C * (1 - R) * Hm
 
     TotalMortality <- 1 - (1 - N) * (1 - U)
     Survivorship <- cumprod(1 - N)
-    Survivorship <- c(1, Survivorship[-n])
+    Survivorship <- c(1, Survivorship[-nt])
     FishedSurvivorship <- cumprod(1 - TotalMortality)
-    FishedSurvivorship <- c(1, FishedSurvivorship[-n])
+    FishedSurvivorship <- c(1, FishedSurvivorship[-nt])
 
     data.frame(Age = t, Length = L, Weight = W, Fecundity = E, Spawning = S,
       NaturalMortality = N, Vulnerability = V, Retention = R,