+++ data-raw/data-raw.R
@@ -1,3 +1,3 @@
 source("data-raw/parameters.R")
-source("data-raw/populations.R")
 source("data-raw/code.R")
+source("data-raw/populations.R")