+++ tests/testthat/test-utils.R
@@ -1,5 +1,5 @@
 context("utils")
 
-test_that("ypr_inst2inter", {
-  expect_equal(ypr_inst2inter(ypr_inter2inst(0.5)), 0.5)
+test_that("inst2inter", {
+  expect_equal(inst2inter(inter2inst(0.5)), 0.5)
 })